Factored zfs_process() out of zfs_call() and zfs_popen()
Could condense these into one call, but it probably makes sense to leave
the logic for choosing between a synchronous call and popen in both send
and receive, where it is useful.  Might still revert to using separate
functions for send / recv returning a file immediately or taking a file
and blocking.