Choose log level based on whether command generates verbose output
Most zfs commands only write to stderr on failure.  Since this will also
generate an exception, it is sufficient to log to DEBUG in this case.  This
also allows failed commands to be silenced in a try/except block with log
levels up to INFO.

When the -v option is set (available for send, receive, destroy, mount, and
upgrade), increase the log level to INFO.