Make subprocess streams consistent between Python 2 & 3
Use automatic buffering (bufsize=-1) by default in both Python 2 & 3,
and always return streams opened with the io module.  The subprocess
module in Python 2 uses open() rather than io.open() and the resulting
streams cannot be wrapped with TextIOWrapper.  This is necessary for
reading text from stderr when stdin or stdout is a binary stream (eg
zfs send and receive).